Struct apple_xar::reader::XarReader

pub struct XarReader<R: Read + Seek + Sized + Debug> { /* private fields */ }

Read-only interface to a single XAR archive.

Implementations

impl<R: Read + Seek + Sized + Debug> XarReader<R>

pub fn new(reader: R) -> XarResult<Self>

Construct a new XAR reader from a stream reader.

pub fn into_inner(self) -> R

Obtain the inner reader.

pub fn header(&self) -> &XarHeader

Obtain the parsed XarHeader file header.

pub fn heap_start_offset(&self) -> u64

The start offset of the heap.

pub fn table_of_contents(&self) -> &TableOfContents

Obtain the table of contents for this archive.

pub fn table_of_contents_decoded_data(&mut self) -> XarResult<Vec<u8>>

Obtain the decoded content of the table of contents.

pub fn checksum_data(&mut self) -> XarResult<Vec<u8>>

Obtain the raw bytes holding the checksum.

pub fn digest_table_of_contents_with( &mut self, checksum: ChecksumType ) -> XarResult<Vec<u8>>

Digest the table of contents content with the specified algorithm.

pub fn files(&self) -> XarResult<Vec<(String, File)>>

Obtain the file entries in this archive.

pub fn find_file(&self, filename: &str) -> XarResult<Option<File>>

Attempt to find the File entry for a given path in the archive.

pub fn write_file_data_heap_from_file( &mut self, file: &File, writer: &mut impl Write ) -> XarResult<usize>

Write heap file data for a given file record to a writer.

This will write the raw data backing a file as stored in the heap. There’s a good chance the raw data is encoded/compressed.

Returns the number of bytes written.

pub fn write_file_data_heap_from_id( &mut self, id: u64, writer: &mut impl Write ) -> XarResult<usize>

Write heap file data for a given file ID to a writer.

This is a wrapper around Self::write_file_data_heap_from_file that resolves the File given a file ID.

pub fn write_file_data_decoded_from_file( &mut self, file: &File, writer: &mut impl Write ) -> XarResult<usize>

Write decoded file data for a given file record to a writer.

This will call Self::write_file_data_heap_from_file and will decode that data stream, if the file data is encoded.

pub fn write_file_data_decoded_from_id( &mut self, id: u64, writer: &mut impl Write ) -> XarResult<usize>

Write decoded file data for a given file ID to a writer.

This is a wrapper for Self::write_file_data_decoded_from_file that locates the File entry given a file ID.

pub fn get_file_data_from_path( &mut self, path: &str ) -> XarResult<Option<Vec<u8>>>

Resolve data for a given path.

pub fn unpack(&mut self, dest_dir: impl AsRef<Path>) -> XarResult<()>

Unpack the contents of the XAR archive to a given directory.

pub fn checksum(&mut self) -> XarResult<(ChecksumType, Vec<u8>)>

Obtain the archive checksum.

The checksum consists of a digest format and a raw digest.

pub fn verify_table_of_contents_checksum(&mut self) -> XarResult<bool>

Validate the recorded checksum of the table of contents matches actual file state.

Will Err if an error occurs obtaining or computing the checksums. Returns Ok with a bool indicating if the checksums matched.

pub fn rsa_signature( &mut self ) -> XarResult<Option<(Vec<u8>, Vec<CapturedX509Certificate>)>>

Obtain RSA signature data from this archive.

The returned tuple contains the raw signature data and the embedded X.509 certificates.

pub fn verify_rsa_checksum_signature(&mut self) -> XarResult<bool>

Verifies the RSA signature in the archive.

This verifies that the RSA signature in the archive, if present, is a valid signature for the archive’s checksum data.

The boolean return value indicates if signature validation was performed.

pub fn cms_signature(&mut self) -> XarResult<Option<SignedData>>

Attempt to resolve a cryptographic message syntax (CMS) signature.

The data signed by the CMS signature is the raw data returned by Self::checksum.

pub fn verify_cms_signature(&mut self) -> XarResult<bool>

Verifies the cryptographic message syntax (CMS) signature, if present.