Struct apple_xar::table_of_contents::TableOfContents

pub struct TableOfContents { /* private fields */ }

An XML table of contents in a XAR file.

Implementations

impl TableOfContents

pub fn from_reader(reader: impl Read) -> XarResult<Self>

Parse XML table of contents from a reader.

pub fn files(&self) -> XarResult<Vec<(String, File)>>

Resolve the complete list of files.

Files are sorted by their numerical ID, which should hopefully also be the order that file data occurs in the heap. Each elements consists of the full filename and the record.

pub fn to_xml(&self) -> XarResult<Vec<u8>>

pub fn write_xml<W: Write>(&self, writer: &mut EventWriter<W>) -> XarResult<()>